+++ src/main/java/com/SoftwareFactoryCustomer/dao/UserDao.java
@@ -1,14 +1,13 @@
 package com.SoftwareFactoryCustomer.dao;
 
+
 import com.SoftwareFactoryCustomer.model.User;
 
 import java.util.List;
 
-
-
 public interface UserDao {
 
-    User findById(int id);
+    User findById(Long id);
 
     User findBySSO(String sso);
