+++ src/main/java/com/SoftwareFactoryCustomer/dao/UserDaoImpl.java
@@ -1,6 +1,5 @@
 package com.SoftwareFactoryCustomer.dao;
 
-import java.util.List;
 
 import com.SoftwareFactoryCustomer.model.User;
 import org.hibernate.Criteria;
@@ -11,15 +10,15 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Repository;
 
-
+import java.util.List;
 
 
 @Repository("userDao")
-public class UserDaoImpl extends AbstractDao<Integer, User> implements UserDao {
+public class UserDaoImpl extends AbstractDao<Long, User> implements UserDao {
 
 	static final Logger logger = LoggerFactory.getLogger(UserDaoImpl.class);
 
-	public User findById(int id) {
+	public User findById(Long id) {
 		User user = getByKey(id);
 		if(user!=null){
 			Hibernate.initialize(user.getUserProfiles());
